Grouped all file-upload functionality.
