+++ template/app/src/server/queries.ts
@@ -1,14 +1,11 @@
-import { type DailyStats, type GptResponse, type User, type PageViewSource, type Task, type File } from 'wasp/entities';
+import { type DailyStats, type GptResponse, type User, type PageViewSource, type Task } from 'wasp/entities';
 import { HttpError } from 'wasp/server';
 import {
   type GetGptResponses,
   type GetDailyStats,
   type GetPaginatedUsers,
   type GetAllTasksByUser,
-  type GetAllFilesByUser,
-  type GetDownloadFileSignedURL,
 } from 'wasp/server/operations';
-import { getDownloadFileSignedURLFromS3 } from './file-upload/s3Utils.js';
 import { type SubscriptionStatusOptions } from '../shared/types.js';
 
 type DailyStatsWithSources = DailyStats & {
@@ -47,29 +44,6 @@ export const getAllTasksByUser: GetAllTasksByUser<void, Task[]> = async (_args, 
       createdAt: 'desc',
     },
   });
-};
-
-export const getAllFilesByUser: GetAllFilesByUser<void, File[]> = async (_args, context) => {
-  if (!context.user) {
-    throw new HttpError(401);
-  }
-  return context.entities.File.findMany({
-    where: {
-      user: {
-        id: context.user.id,
-      },
-    },
-    orderBy: {
-      createdAt: 'desc',
-    },
-  });
-};
-
-export const getDownloadFileSignedURL: GetDownloadFileSignedURL<{ key: string }, string> = async (
-  { key },
-  _context
-) => {
-  return await getDownloadFileSignedURLFromS3({ key });
 };
 
 export const getDailyStats: GetDailyStats<void, DailyStatsValues> = async (_args, context) => {
